backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pymongo==4.7.2
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
requests==2.31.0
